Updated database constraints and added creator checks for forums endpoints

+++ backend/controllers/forumController.js
@@ -33,6 +33,11 @@ exports.getForumByIdController = async (req, res) => {
     try {
         console.log('Fetching forum with ID: ', req.params.forumID);
         const forum = await forumModel.getForumByIdModel(req.params.forumID);
+
+        if(!forum) {
+            console.log('Forum not found');
+            return res.status(404).json({ message: 'Forum not found' });
+        }
         res.status(200).json(forum);
     } catch (error) {
         res.status(500).json({ message: 'Error fetching forum' });
@@ -44,6 +49,14 @@ exports.getForumByIdController = async (req, res) => {
 // TODO: Set it to check user creator of Forum. Maybe make it to where only the forum description can be changed. if not make checks for if part of form is null/blank
 exports.updateforumController = async (req, res) => {
     try {
+        const userID = req.user.id;
+        const oldForum = await forumModel.getForumByIdModel(req.params.forumId);
+
+        if(oldForum.created_by != userID) {
+            console.log('Not authorized to update forum');
+            return res.status(401).json({ message: 'Not authorized to update forum' });
+        }
+
         console.log('Updating forum with ID: ', req.params.forumId, ' with data: ', req.body);
         const updatedForum = await forumModel.updateForumModel(req.params.forumId, req.body);
         res.status(200).json(updatedForum);
@@ -56,6 +69,16 @@ exports.updateforumController = async (req, res) => {
 // Delete forum by ID
 exports.deleteforumController = async (req, res) => {
     try {
+        const userID = req.user.id;
+        const oldForum = await forumModel.getForumByIdModel(req.params.forumId);
+        console.log(userID)
+        console.log(oldForum.created_by)
+
+        if(oldForum.created_by != userID) {
+            console.log('Not authorized to delete forum');
+            return res.status(401).json({ message: 'Not authorized to delete forum' });
+        }
+
         console.log('Deleting forum with ID: ', req.params.forumId);
         const deletedForum = await forumModel.deleteForumModel(req.params.forumId);
         res.status(200).json(deletedForum);
